Different approach: single structure as a global
Did away with the multiple moving parts approach that seemed good at first but got uglier fast.  Now a single list of all current values is a global.  I will make an update function that updates it whenever it is changed so only one function ever modifies it.

I ended up keeping the grid (before I had rows, columns, and grids as data structures) because it's necessary information for solving the puzzle.  Row and column can be obtained with a bit of math based on the current index, but grid was not so straightforward.  So I made a few dictionaries so I could look them up whenever I needed to (may only end up needing the reverse dictionary, but I made both just in case).

Happy New Year, I guess.